+++ src/time.rs
@@ -6,17 +6,18 @@ pub trait DurationSinceEpoch {
     fn as_secs(&self) -> u64;
 }
 
-use core::time::Duration;
-use std::time::SystemTime;
-
+#[cfg(feature = "std")]
 /// Duration since the epoch represented by standard library types.
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, PartialOrd, Ord)]
-pub struct StdDurationSinceEpoch(Duration);
+pub struct StdDurationSinceEpoch(core::time::Duration);
 
+#[cfg(feature = "std")]
 impl StdDurationSinceEpoch {
     /// Returns the current duration since the epoch.
     #[must_use]
     pub fn now() -> Self {
+        use std::time::SystemTime;
+
         StdDurationSinceEpoch(
             SystemTime::now()
                 .duration_since(SystemTime::UNIX_EPOCH)
@@ -25,6 +26,7 @@ impl StdDurationSinceEpoch {
     }
 }
 
+#[cfg(feature = "std")]
 impl DurationSinceEpoch for StdDurationSinceEpoch {
     fn as_secs(&self) -> u64 {
         self.0.as_secs()