+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add `no_std` functionality
+
 ## [0.4.0] - 2022-05-03
 
 ### Added
@@ -48,7 +54,7 @@
 
 - Initial implementation.
 
-[Unreleased]: https://github.com/bluk/gen_value/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/bluk/min_jwt/compare/v0.4.0...HEAD
 [0.4.0]: https://github.com/bluk/min_jwt/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/bluk/min_jwt/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/bluk/min_jwt/compare/v0.2.0...v0.2.1